fix(reframed): improve the document.body virtualization
Previously we returned the first child of the shadowroot, which can result in head being returned.

Now we first look for a special element called wf-body, and if not found, we return the first child that has any children already.

This should still be made more robust because for example if the head element is preserved but renamed to something else, we'll likely start treating it as a body element which is not good.